This is data science project for health care
sample project using basic statistical analysis
Outcomes of this project are
Download a DataSet from *.xlsx files
Automatically change data in the DataSet
Transform the table
Visualize the data with pandas and seaborn
Expect
Minimum and maximum value
Average
Quarters
Pivot tables


In this project, I learned the methods of statistical analysis of data obtained via a people survey on the disease of COVID-19. The laboratory consists of four stages:

Data download
Data preparation
Statistical analysis
Data visualization
The first stage will show you how to download spreadsheet data:

indexing of key fields
working with data types such as datatime
writing a parser function to automatically create index fields


At the next stage, it will be demonstrated how to pre-prepare data for analysis, in particular:

elimination of missing data
replacement and change of cell data
change the data types of columns


At the stage of statistical analysis, the methods of automation of statistical information calculation will be demonstrated, in particular:

calculation of basic statistical indicators
data sorting and grouping
data transformation DataSet
calculation of statistical indicators on the transformed data


At the stage of visualization, the possibilities of visualization of the received statistical indicators are shown.